% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_percent.R
\name{as_percent}
\alias{as_percent}
\title{As percent}
\usage{
as_percent(x, round = 0)
}
\arguments{
\item{x}{A real number (typically between 0 and 1)}

\item{round}{An integer of how many digits to round to}
}
\value{
a string
}
\description{
Convert a real number to a string percentage for output
}
