% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_grouped.R, R/barchart_ungrouped.R
\name{barchart_grouped}
\alias{barchart_grouped}
\alias{barchart_ungrouped}
\title{Bar chart}
\usage{
barchart_grouped(
  .data,
  bar_field,
  grouping_field = NULL,
  xaxis_title = "",
  yaxis_title = "",
  grouping_mode = c("stack", "group"),
  legend_title = "",
  plot_title = ""
)

barchart_ungrouped(
  .data,
  bar_field,
  xaxis_title = "",
  yaxis_title = "",
  bar_color = "#26547C",
  height = NA_real_,
  legend_title = "",
  plot_title = "",
  sort_method = c("desc", "alphabetical", "numeric")
)
}
\arguments{
\item{.data}{A data frame}

\item{bar_field}{The column to use as the bar_chart basis}

\item{grouping_field}{The column in \code{.data} to group the bar chart by}

\item{xaxis_title}{A character vector}

\item{yaxis_title}{A character vector}

\item{grouping_mode}{Either "stack" or "group"}

\item{legend_title}{A character vector}

\item{plot_title}{A character vector}

\item{bar_color}{A character vector of a hex color, or similarly handled color in Plotly.}

\item{height}{A real number to define the height of the plot (in pixels)}

\item{sort_method}{See \code{add_sorting}}
}
\description{
Create a bar chart, either grouped by a \code{grouping_field}, or ungrouped.
}
\section{Functions}{
\itemize{
\item \code{barchart_grouped()}: Grouped bar chart

}}
