% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sas_dir_to_arrow.R
\name{convert_sas_dir_to_arrow}
\alias{convert_sas_dir_to_arrow}
\title{Convert SAS data in directory to arrow}
\usage{
convert_sas_dir_to_arrow(...)
}
\arguments{
\item{...}{strings to be joined together into a path. (Can contain slashes but not needed)}
}
\description{
Converts all SAS files (.sas7bdat) in a diretory to feather files (.arrow) in the same directory
}
