% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sas_file_to_arrow.R
\name{convert_sas_file_to_arrow}
\alias{convert_sas_file_to_arrow}
\title{Convert SAS file to arrow}
\usage{
convert_sas_file_to_arrow(...)
}
\arguments{
\item{...}{strings to be joined together into a path. (Can contain slashes but not needed)}
}
\description{
Converts a SAS file (.sas7bdat) to a feather file (.arrow) in the same directory
}
