% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_copy.R
\name{create_directory_for_file}
\alias{create_directory_for_file}
\title{Create directories}
\usage{
create_directory_for_file(file)
}
\arguments{
\item{file}{a single string or a vector of strings representing file names}
}
\description{
Creates local directories (if needed) based on file path(s)
}
