% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_unique.R
\name{df_confirm_unique}
\alias{df_confirm_unique}
\title{Confirm unique}
\usage{
df_confirm_unique(.data, ...)
}
\arguments{
\item{.data}{a data frame to be tested}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Columns representing the primary key of the data frame}
}
\description{
Returns an error if not unique. Otherwise, returns the originating data frame
}
