% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_histogram.R
\name{df_histogram}
\alias{df_histogram}
\title{Quick histogram}
\usage{
df_histogram(.data, .col)
}
\arguments{
\item{.data}{A data frame containing the column \code{.col}}

\item{.col}{The column for which to create a histogram}
}
\description{
Quick histogram
}
