% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_kableize.R
\name{df_kableize}
\alias{df_kableize}
\title{Kableize a data frame}
\usage{
df_kableize(t, var_col_name, val_col_name = "All subjects")
}
\arguments{
\item{t}{A data frame}

\item{var_col_name}{The desired output name of the first column}

\item{val_col_name}{The desired output name of the second column (which will be appended with "(N={n_pts})")}
}
\description{
This function takes a data frame with a first column of row headers and the second column as a cohort of
patients. The first row, "n_pts", represents the number of patients in each cohort.
}
\details{
The function builds a kable object, appending "(N={n_pts})". It optionally adds indenting based on an
\code{indent} column.
}
