% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_piechart.R
\name{df_piechart}
\alias{df_piechart}
\title{Quick pie chart}
\usage{
df_piechart(
  .data,
  pie_field,
  plot_title = "",
  legend_title = "",
  palette = station_palette
)
}
\arguments{
\item{.data}{A data frame}

\item{pie_field}{The (unquoted) field in .data to create pie chart from}

\item{plot_title}{A character vector}

\item{legend_title}{A character vector}

\item{palette}{A palette of colors}
}
\description{
Quick pie chart
}
