% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_body_nodeset.R, R/docx_xml.R
\name{docx_body_nodeset}
\alias{docx_body_nodeset}
\alias{docx_body_nodeset.rdocx}
\alias{docx_body_nodeset.xml_document}
\alias{docx_xml}
\title{Get the body nodeset of an rdocx or rdocx XML element}
\usage{
docx_body_nodeset(x)

\method{docx_body_nodeset}{rdocx}(x)

\method{docx_body_nodeset}{xml_document}(x)

docx_xml(r)
}
\arguments{
\item{x}{An officer::rdocx or an xml2::xml_document object}

\item{r}{An officer::rdocx object}
}
\description{
Get the body nodeset of an rdocx or rdocx XML element
}
\section{Functions}{
\itemize{
\item \code{docx_xml()}: Get the xml2::xml_document object from an officer::rdocx object

}}
