% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_extract_table.R
\name{docx_extract_table}
\alias{docx_extract_table}
\title{Extract a table from a officer::rdocx object}
\usage{
docx_extract_table(r, table_num = 1)
}
\description{
Extract a table from a officer::rdocx object
}
