% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_insert_table.R
\name{docx_insert_table}
\alias{docx_insert_table}
\title{Insert table into docx}
\usage{
docx_insert_table(
  r,
  df,
  header_text = NULL,
  header_num = NULL,
  table_title = NULL,
  align_table = NULL,
  source_text = NULL,
  footnote_text = NULL,
  node_details = NULL,
  toc_tlf_num = NULL
)
}
\arguments{
\item{r}{an officer::rdocx object}

\item{df}{a data frame}

\item{header_text}{The regex to search for (for inserting the text)}

\item{header_num}{The header number}

\item{table_title}{If not \code{NULL}, insert a table title above the table}
}
\description{
Insert table into docx
}
