% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_insert_text.R
\name{docx_insert_text}
\alias{docx_insert_text}
\title{Insert table into docx}
\usage{
docx_insert_text(
  r,
  text_to_insert,
  header_text,
  header_num,
  table_title = NULL,
  node_details = NULL,
  align_table = NULL,
  footnote_text = NULL
)
}
\arguments{
\item{r}{an officer::rdocx object}

\item{header_text}{The regex to search for (for inserting the text)}

\item{header_num}{The header number}

\item{table_title}{If not \code{NULL}, insert a table title above the table}

\item{df}{a data frame}
}
\description{
Insert table into docx
}
