% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_node_details.R
\name{docx_node_details}
\alias{docx_node_details}
\title{Get the node details from an officer::rdocx or an xml2::xml_document object}
\usage{
docx_node_details(x)
}
\arguments{
\item{x}{An officer::rdocx object or an xml2::xml_document object}
}
\description{
Extract a data frame of nodes from the body of an officer::rdocx or an xml2::xml_document object
}
