% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_remove_paragraph.R
\name{docx_remove_paragraph}
\alias{docx_remove_paragraph}
\title{Remove a paragraph found from a regular expression}
\usage{
docx_remove_paragraph(r, location_regex)
}
\arguments{
\item{r}{an officer::rdocx object}

\item{location_regex}{The regex to search for (for removing the paragraph)}
}
\description{
Remove a paragraph found from a regular expression
}
