% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_remove_text.R
\name{docx_remove_text}
\alias{docx_remove_text}
\title{Remove text (docx)}
\usage{
docx_remove_text(r, location_regex)
}
\arguments{
\item{r}{an officer::rdocx object}

\item{location_regex}{The regex to search for (for inserting the text)}
}
\description{
Remove text (docx)
}
