% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_replace_figure.R
\name{docx_replace_figure}
\alias{docx_replace_figure}
\title{Insert figure into a docx}
\usage{
docx_replace_figure(
  r,
  fig_path,
  caption_text,
  style = NULL,
  node_details = NULL
)
}
\arguments{
\item{r}{an officer::rdocx object}

\item{fig_path}{The relative path of the figure to insert}

\item{caption_text}{The text to search for (for inserting the figure)

#' @details The index of the figure is discovered by looking for the captions beyond the Table of Contents on
the document. The \code{caption_text} is searched (via regex), and a table is searched for between the following
nodes and the next caption. If exactly one is found, it is replaced with df. Otherwise, an error is
thrown.}
}
\value{
an altered rdocx object
}
\description{
Insert figure into a docx
}
