% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_replace_table.R
\name{docx_replace_table}
\alias{docx_replace_table}
\title{Replace a table within an officer::rdocx object}
\usage{
docx_replace_table(r, df, caption_text, node_details = NULL)
}
\arguments{
\item{r}{An rdocx object}

\item{df}{A data frame to be inserted}

\item{caption_text}{The location to insert the table. See @details.}

\item{node_details}{Optional. A data frame returned from \code{node_details()}. If \code{NULL}, it will be
generated.}
}
\description{
Replace a table within an officer::rdocx object
}
\details{
The index of the table is discovered by looking for the captions beyond the Table of Contents on
the document. The \code{caption_text} is searched (via regex), and a table is searched for between the following
nodes and the next caption. If exactly one is found, it is replaced with df. Otherwise, an error is
thrown.

Table properties from the original table are kept.
}
