% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_replace_text.R
\name{docx_replace_text}
\alias{docx_replace_text}
\title{Replace text within an rdocx}
\usage{
docx_replace_text(r, new_text, text_to_replace, node_details = NULL)
}
\arguments{
\item{r}{An officer::rdocx object}

\item{new_text}{Text that should be inserted}

\item{text_to_replace}{Text that should be replaced}

\item{node_details}{A node_details data frame from docx_node_details(). If \code{NULL}, this will be derived.}

\item{caption}{Document caption in which to search for the given text. If \code{NULL}, search the whole document.}

\item{instance}{Which instance of the text to replace. If \code{NULL}, replace all.}
}
\description{
Replace text within an rdocx
}
