% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_get_dim.R
\name{figure_dimensions}
\alias{figure_dimensions}
\alias{pixel_to_inch}
\alias{fig_get_dim_emf}
\alias{raw_to_32bit_unsigned}
\title{Get figure dimensions}
\usage{
figure_dimensions(fig_path = file.path(R.home("doc"), "html", "logo.jpg"))

pixel_to_inch(pixel, ppi = 96)

fig_get_dim_emf(fig_path)

raw_to_32bit_unsigned(raw_32bit)
}
\arguments{
\item{fig_path}{The relative path of the figure to insert}

\item{pixel}{The pixel count, returned from a dim() call}

\item{ppi}{The pixels per inch to use. The usual pixel count of monitors is 94 and 120 pixels per inch
(PPI). One site said the most common is 96, so we use that as the default.}

\item{raw_32bit}{A raw vector of length 32 bits (4 bytes)}
}
\value{
a named real number vector
}
\description{
Get figure dimensions
}
\section{Functions}{
\itemize{
\item \code{pixel_to_inch()}: Convert pixels to inches

\item \code{fig_get_dim_emf()}: Get EMF file dimensions

\item \code{raw_to_32bit_unsigned()}: Get 32-bit unsigned value

}}
