% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_copy.R
\name{file_copy_to}
\alias{file_copy_to}
\title{NavigatR file copy}
\usage{
file_copy_to(...)
}
\arguments{
\item{...}{see \link[fs]{file_copy}}
}
\description{
Wrapper around fs::file_copy to fs::dir_create() if necessary
}
