% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_create_directory.R
\name{file_create_directory}
\alias{file_create_directory}
\title{Create directories}
\usage{
file_create_directory(file)
}
\arguments{
\item{file}{a single string or a vector of strings representing file names}
}
\description{
Creates local directories (if needed) based on file path(s)
}
