% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_copy.R, R/file_get_directory_path.R
\name{file_get_directory_path}
\alias{file_get_directory_path}
\title{Get file directory}
\usage{
file_get_directory_path(p)

file_get_directory_path(p)
}
\arguments{
\item{p}{a character vector representing a file path}
}
\description{
Get the file directory given a full file path

Get the file directory given a full file path
}
