% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_get_name.R
\name{file_get_name}
\alias{file_get_name}
\title{Get file name}
\usage{
file_get_name(p)
}
\arguments{
\item{p}{a character vector representing a file path}
}
\description{
Get the file name from a file path
}
