% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_get_suffix.R
\name{file_get_suffix}
\alias{file_get_suffix}
\title{Get file suffix}
\usage{
file_get_suffix(p)
}
\arguments{
\item{p}{a character vector representing a file name or file path}
}
\description{
Get the suffix of a file name or file path
}
