% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_elser.R
\name{if_elser}
\alias{if_elser}
\title{If/Else statement on one line}
\usage{
if_elser(test, yes, no)
}
\arguments{
\item{test}{A logical}

\item{yes}{Output if \code{test} is TRUE}

\item{no}{Output if \code{test} is FALSE}
}
\description{
If/Else statement on one line
}
