% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R, R/util.R
\name{list_files}
\alias{list_files}
\title{List the files in a directory}
\usage{
list_files(directory_path = getwd(), file_regex = ".*", recursive = FALSE)

list_files(directory_path = getwd(), file_regex = ".*", recursive = FALSE)
}
\arguments{
\item{directory_path}{A string representing a directory (local or remote)}

\item{file_regex}{A string representing a regex to filter to}

\item{directory}{A string representing a directory (local or remote)}

\item{sub_dir}{An optional subdirectory. This is a separate parameter so we can loop over directories
and access the same subdirectories (built for V:/.../sp/outputs/)}
}
\description{
This function automatically removes temporary files (i.e. those with '~' in the path)

This function automatically removes temporary files (i.e. those with '~' in the path)
}
