% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{name_and_title_df}
\alias{name_and_title_df}
\title{Name/Title compounds, studies, and tasks in a data frame}
\usage{
name_and_title_df(df)
}
\arguments{
\item{df}{A data frame with compound, study, and task columns}
}
\description{
Name/Title compounds, studies, and tasks in a data frame
}
