% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_detect.R
\name{node_detect}
\alias{node_detect}
\title{Test whether a node matches a pattern in its text}
\usage{
node_detect(n, text, fixed = F, ignore.case = F)
}
\arguments{
\item{n}{A node or nodeset}

\item{text}{Text to search for. See @details for more information}

\item{fixed}{Logical. See \code{\link[=grepl]{grepl()}}}

\item{ignore.case}{Logical. See \code{\link[=grepl]{grepl()}}}
}
\description{
Test whether a node matches a pattern in its text
}
\details{
The search is conducted only on the alphanumeric characters. Any non-alphanumeric characters are replaced
with an any-character '.' search. This is to correct for some seemingly odd characters in the template
we see for 877, e.g. 'Figure.1' works, but 'Figure 1' does not.
}
