% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_node_details.R
\name{node_is_caption}
\alias{node_is_caption}
\title{Determine whether an xml node is a caption.}
\usage{
node_is_caption(n)
}
\arguments{
\item{n}{A node or nodeset}
}
\description{
Determine whether an xml node is a caption.
}
