% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_node_details.R
\name{node_is_header}
\alias{node_is_header}
\title{Determine whether an xml node is a header. Takes in a node or a nodeset, and returns T/F}
\usage{
node_is_header(n, header_num = NULL)
}
\arguments{
\item{n}{A node or nodeset}

\item{header_num}{An integer if we want to filter to a header of a certain type}
}
\description{
Determine whether an xml node is a header. Takes in a node or a nodeset, and returns T/F
}
