% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_node_details.R
\name{node_is_style}
\alias{node_is_style}
\title{Determine whether an xml node has a specified style.}
\usage{
node_is_style(n, pattern)
}
\arguments{
\item{n}{A node or nodeset}
}
\description{
Determine whether an xml node has a specified style.
}
