% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nv_file_copy}
\alias{nv_file_copy}
\title{NavigatR file copy}
\usage{
nv_file_copy(...)
}
\arguments{
\item{...}{see \link[fs]{file_copy}}
}
\description{
Wrapper around fs::file_copy to fs::dir_create() if necessary
}
