% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_add_lines.R
\name{plotly_add_lines}
\alias{plotly_add_lines}
\title{Add lines}
\usage{
plotly_add_lines(
  p,
  ref_val = c(-0.8, -0.3, 0),
  color = "#26547C",
  text = "ULN",
  dash_type = "dot",
  refaxis = "y",
  linetype = "horizontal"
)
}
\arguments{
\item{p}{a plotly object}

\item{ref_val}{The locations om the \code{refaxis} to draw the line (an x-axis or y-axis value)}

\item{color}{The color of the line}

\item{text}{The text to display for the line}

\item{dash_type}{Character vector passed to plotly \code{shapes$line$dash}}

\item{refaxis}{Use "x" or "y", or if there are multiple y-axes, the trace yaxis specified}

\item{linetype}{'horizontal' or 'vertical'}
}
\description{
Add lines to a plot. Unfortunately, this can only be called once per plotly::plot_ly(), so gather the
line requirements prior to call. Most of the parameters are extensible for multiple lines.
}
