% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_custom_layout.R
\name{plotly_custom_layout}
\alias{plotly_custom_layout}
\title{WebPro Plotly Layout}
\usage{
plotly_custom_layout(
  .p,
  xaxis_title = "",
  yaxis_title = "",
  showgrid = F,
  showzeroline = F,
  legend_orientation = "h"
)
}
\arguments{
\item{.p}{a Plotly object}

\item{xaxis_title}{a character vector}

\item{yaxis_title}{a character vector}
}
\description{
A custom to layout to apply to Plotly objects
}
