% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_insert_image.R
\name{pptx_insert_figure}
\alias{pptx_insert_figure}
\title{Insert a figure into a officer::rpptx object}
\usage{
pptx_insert_figure(x, fig_path, location_regex)
}
\arguments{
\item{x}{An officer::rpptx object}

\item{fig_path}{Path to the figure to be inserted}

\item{location_regex}{The regex to search for within \code{x} where to insert the figure}
}
\description{
Insert a figure into a officer::rpptx object
}
