% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feather.R
\name{read_feather}
\alias{read_feather}
\title{Read saved file}
\usage{
read_feather(...)
}
\arguments{
\item{...}{strings to be pasted together to form a path (via \code{fs::path(...)})}
}
\description{
Read a saved feather file
}
