% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_get_dir.R
\name{remote_get_dir}
\alias{remote_get_dir}
\title{Get remote directory}
\usage{
remote_get_dir(compound = NULL, study = NULL, task = NULL)
}
\arguments{
\item{compound}{The compound of the study (directory in V:/). If NULL, infer from config.}

\item{study}{The study (directory in V:/\link{directory}). If NULL, infer from config.}

\item{task}{The study (directory in V:/\link{directory}). If NULL, infer from config.}
}
\description{
Get the remote (V:/) directory of interest
}
