% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_pull_data.R
\name{remote_pull_data}
\alias{remote_pull_data}
\title{Pull remote data}
\usage{
remote_pull_data(
  remote_files_of_interest,
  local_data_directory = get_golem_config("local_data_directory"),
  overwrite = T
)
}
\arguments{
\item{remote_files_of_interest}{A list of vectors of files. See Details.}

\item{local_data_directory}{The local directory in which to create a timestamped directory containing the data.}

\item{overwrite}{Logical. If \code{TRUE}, overwrite existing files.}

\item{remote_dir}{The study directory on the remote server.}

\item{sub_dir}{The specific subdirectories of the remote path to search. For example, if the
\code{remote_path} is V:/compound/study and the \code{sub_dir} is '/sp/outputs', the paths of
V:/compound/study/.../sp/outputs/ will be searched.}
}
\description{
Pulls remote data as-is into \code{local_data_directory}. This function pulls the \code{files_to_pull} over into
a timestamped directory (with today's date).
}
\details{
If the directory/data exists, the function will not delete any files, but will replace any
existing data.

Note that \code{remote_files_of_interest} should have the following structure:
\itemize{
\item{Remote}{ \code{remote_dir}/\code{names(tasks_and_files)}/\code{remote_sub}/\code{values(tasks_and_files)} }
\item{Local}{ ~/\code{local_data_directory}/\code{names(tasks_and_files)}/\code{values(tasks_and_files)} }
}
}
