% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_combine.R
\name{rtf_combine}
\alias{rtf_combine}
\title{This is heavily inspired by https://www.lexjansen.com/pharmasug/2010/PO/PO05.pdf}
\usage{
rtf_combine(files, output_path)
}
\description{
This is heavily inspired by https://www.lexjansen.com/pharmasug/2010/PO/PO05.pdf
}
