% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_convert_to_pdf_here.R
\name{rtf_convert_to_pdf_here}
\alias{rtf_convert_to_pdf_here}
\title{Convert files to PDFs}
\usage{
rtf_convert_to_pdf_here(file)
}
\arguments{
\item{file}{a single string or a vector of strings representing file names}
}
\description{
Convert local files to PDFs (in the same directory)
}
