% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_extract_img.R
\name{rtf_extract_figure}
\alias{rtf_extract_figure}
\title{Extract figure from rtf file and return its location}
\usage{
rtf_extract_figure(file_name, which_fig = NULL)
}
\arguments{
\item{file_name}{rtf file location}

\item{which_fig}{Which figure to pull (experimental)}
}
\description{
For now, we assume there is exactly one media file
}
