% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_extract_footnote.R
\name{rtf_extract_footnote}
\alias{rtf_extract_footnote}
\title{Extract the footnote from an RTF file}
\usage{
rtf_extract_footnote(file_path)
}
\arguments{
\item{file_path}{Path to an RTF file}
}
\description{
See
}
