% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_extract_table.R
\name{rtf_extract_tables}
\alias{rtf_extract_tables}
\title{Extract a table (data frame) from an RTF file}
\usage{
rtf_extract_tables(file_path, format_cells = T)
}
\arguments{
\item{file_path}{The path to the RTF file}
}
\description{
Extract a table (data frame) from an RTF file
}
