% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_get_title.R
\name{rtf_get_title}
\alias{rtf_get_title}
\title{Get RTF file title}
\usage{
rtf_get_title(p, strip_non_alphanumeric = F)
}
\arguments{
\item{p}{path to the file (character vector)}
}
\description{
Get the first text in the RTF file, remove '|' and '*', and trim whitespace
}
