% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_reformat.R
\name{rtf_reformat}
\alias{rtf_reformat}
\title{Retitle RTF file}
\usage{
rtf_reformat(file, new_title = NULL)
}
\arguments{
\item{file}{a single string or a vector of strings representing file names}

\item{new_title}{}
}
\description{
RTF files contain a title that gets displayed at the top of the PDF file (when converted). This function
retitles the RTF file.
}
