% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_list_bucket_contents.R
\name{s3_list_bucket_contents}
\alias{s3_list_bucket_contents}
\title{List S3 bucket contents}
\usage{
s3_list_bucket_contents(bucket, file_regex = ".*\\\\.rtf$")
}
\arguments{
\item{bucket}{The bucket on S3 to list contents of}

\item{file_regex}{The regex to apply to subset the file list}
}
\description{
List S3 bucket contents
}
