% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_pull_bucket.R
\name{s3_pull_bucket}
\alias{s3_pull_bucket}
\title{List S3 bucket contents}
\usage{
s3_pull_bucket(bucket, local_data_directory, file_regex = ".*\\\\.rtf$")
}
\arguments{
\item{bucket}{The bucket on S3 to pull contents of}

\item{local_data_directory}{The local directory to pull the bucket to}

\item{file_regex}{The regex to apply to subset the file list}
}
\description{
List S3 bucket contents
}
