% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_pull_file.R
\name{s3_pull_file}
\alias{s3_pull_file}
\title{Pull S3 file}
\usage{
s3_pull_file(object, bucket, local_data_directory)
}
\arguments{
\item{object}{The object (file) to pull from S3}

\item{bucket}{The bucket on S3 to pull contents of}

\item{local_data_directory}{The local directory to pull the bucket to}
}
\description{
Pull S3 file
}
