% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_push_local_file.R, R/s3_push_local_files.R
\name{s3_push_local_file}
\alias{s3_push_local_file}
\alias{s3_push_local_files}
\title{Push a single file to S3}
\usage{
s3_push_local_file(file, bucket, local_data_directory)

s3_push_local_files(
  files = list.files(local_data_directory, recursive = T),
  local_data_directory
)
}
\arguments{
\item{file}{The local file to push. Note that the path on S3 will mimic the path from the \code{local_data_directory}}

\item{bucket}{The bucket to push it to}

\item{local_data_directory}{the local directory the file is located in.}

\item{files}{A list of files to push}
}
\description{
Push a single file to S3
}
\section{Functions}{
\itemize{
\item \code{s3_push_local_files()}: 

}}
