% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas_extract_table.R
\name{sas_extract_table}
\alias{sas_extract_table}
\title{Extract SAS table}
\usage{
sas_extract_table(sas_path)
}
\arguments{
\item{sas_path}{The path to the .sas7bdat file}
}
\description{
Extract SAS table
}
