% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_feather.R
\name{save_feather}
\alias{save_feather}
\title{Save feather}
\usage{
save_feather(x, ...)
}
\arguments{
\item{x}{a data frame to save}

\item{...}{strings to be pasted together to form a path (via \code{fs::path(...)})}
}
\description{
Save a file as a feather file
}
