% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rds.R
\name{save_rds}
\alias{save_rds}
\title{Save RDS}
\usage{
save_rds(x, ...)
}
\arguments{
\item{x}{a data frame to save}

\item{...}{strings to be pasted together to form a path (via \code{fs::path(...)})}
}
\description{
Save a file as am RDS file
}
