% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_sas.R
\name{save_sas}
\alias{save_sas}
\title{Save SAS}
\usage{
save_sas(x, ...)
}
\arguments{
\item{x}{a data frame to save}

\item{...}{strings to be pasted together to form a path (via \code{fs::path(...)})}
}
\description{
Save a file as a SAS file
}
