% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_xpt.R
\name{save_xpt}
\alias{save_xpt}
\title{Save XPT}
\usage{
save_xpt(x, ...)
}
\arguments{
\item{x}{a data frame to save}

\item{...}{strings to be pasted together to form a path (via \code{fs::path(...)})}
}
\description{
Save a file as an XPT file
}
