% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_extract_table.R
\name{separate_and_title}
\alias{separate_and_title}
\title{Separate and title a data frame with character vector \code{value}}
\usage{
separate_and_title(df, column_sep = " \\\\| ")
}
\arguments{
\item{df}{A data frame}

\item{column_sep}{Regex to split the data frame by}
}
\description{
Sets the names of the data frame and separates the \code{value} column into columns. Assumes the titles are
in the first row
}
