% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_get.R, R/util.R
\name{sftp_get}
\alias{sftp_get}
\title{SFTP get a remote directory in a local directory}
\usage{
sftp_get(remote_directory, local_data_directory)

sftp_get(remote_directory, local_data_directory)
}
\arguments{
\item{remote_directory}{The remote path to transfer the data to}

\item{local_data_directory}{A local directory containing data to transfer to a remote server}
}
\description{
Works in conjunction with the \code{config} package
}
