% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_put.R, R/util.R
\name{sftp_put}
\alias{sftp_put}
\title{SFTP put a local directory to a remote directory}
\usage{
sftp_put(local_data_directory, remote_directory)

sftp_put(local_data_directory, remote_directory)
}
\arguments{
\item{local_data_directory}{A local directory containing data to transfer to a remote server}

\item{remote_directory}{The remote path to transfer the data to}
}
\description{
Works in conjunction with the \code{config} package
}
