% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_command.R, R/util.R
\name{ssh_command}
\alias{ssh_command}
\title{SSH command}
\usage{
ssh_command(command)

ssh_command(command)
}
\arguments{
\item{command}{A string}
}
\description{
Used in conjunction with the \code{config} package. Builds and calls an ssh call to run a command.
}
