% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station-package.R
\docType{package}
\name{station-package}
\alias{station}
\alias{station-package}
\title{station: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
