% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_indent_DT.R
\name{str_indent_DT}
\alias{str_indent_DT}
\title{Indent string}
\usage{
str_indent_DT(x)
}
\arguments{
\item{x}{a string}
}
\description{
Use in conjunction with DT with \code{escape = FALSE}
}
\details{
https://github.com/rstudio/DT/issues/215
}
