% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_locate_last.R, R/util.R
\name{str_locate_last}
\alias{str_locate_last}
\title{Locate the last location of a character in a string}
\usage{
str_locate_last(string, pattern)

str_locate_last(string, pattern)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible to one.}

\item{pattern}{Pattern (Character) to look for.}
}
\description{
See \link[stringr]{str_sub}

See \link[stringr]{str_sub}
}
