% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_prefix.R, R/util.R
\name{str_prefix}
\alias{str_prefix}
\title{Prefix string}
\usage{
str_prefix(string, ..., sep = "")

str_prefix(string, ..., sep = "")
}
\arguments{
\item{string}{a string}

\item{...}{see \link[stringr]{str_c}}

\item{sep}{see \link[stringr]{str_c}}
}
\description{
Used for piping: apply the following parameters prior to the first

Used for piping: apply the following parameters prior to the first
}
