% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{str_remove}
\alias{str_remove}
\title{Prefix string}
\usage{
str_remove(string, pattern)
}
\arguments{
\item{string}{A string}

\item{pattern}{Pattern to remove. See \link[stringr]{str_replace}}
}
\description{
Used for piping: apply the following parameters prior to the first
}
