% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_format_cells.R
\name{table_format_cells}
\alias{table_format_cells}
\title{Remove spaces in parentheses throughout a table}
\usage{
table_format_cells(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
Many of the RTF files have formatting like '  21 ( 0.0)'. This function removes leading spaces within
the parentheses in every cell: '  21 (0.0)'
}
