% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_node.R
\name{table_to_node}
\alias{table_to_node}
\title{Convert a data frame to a Microsoft Word table}
\usage{
table_to_node(
  df,
  preserve_spacing = F,
  indent_col1 = T,
  style = "CSR-Table",
  pos = "after",
  header = TRUE,
  alignment = NULL,
  align_table = "center",
  stylenames = officer::table_stylenames(rlang::set_names(rep("C-Table Text", ncol(df)),
    nm = colnames(df))),
  first_row = TRUE,
  first_column = TRUE,
  last_row = FALSE,
  last_column = FALSE,
  no_hband = FALSE,
  no_vband = TRUE,
  escape_cells = TRUE
)
}
\arguments{
\item{df}{A data frame}

\item{...}{Parameters passed to \code{officer} functions}
}
\description{
Convert a data frame to a Microsoft Word table
}
