% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_extract_table.R
\name{table_use_first_row_as_names}
\alias{table_use_first_row_as_names}
\title{Use the first row as column names for a data frame}
\usage{
table_use_first_row_as_names(df)
}
\description{
Use the first row as column names for a data frame
}
