% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablenode_indent.R
\name{tablenode_indent}
\alias{tablenode_indent}
\title{Set the formatting to indent a table node when it's the first column and it starts with two spaces}
\usage{
tablenode_indent(t)
}
\description{
Set the formatting to indent a table node when it's the first column and it starts with two spaces
}
