% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_node.R
\name{text_escape_chars}
\alias{text_escape_chars}
\title{From officer::R/utils.R. Escape special characters.}
\usage{
text_escape_chars(text, attribute = FALSE)
}
\description{
From officer::R/utils.R. Escape special characters.
}
