% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_node.R
\name{text_to_node}
\alias{text_to_node}
\title{Convert text to a Microsoft Word paragraph}
\usage{
text_to_node(t, style = "Normal", bold = F, r = NULL)
}
\arguments{
\item{t}{A character vector}

\item{bold}{Logical. Whether to bold the font or not.}
}
\description{
Convert text to a Microsoft Word paragraph
}
