% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unbold_if_clicked}
\alias{unbold_if_clicked}
\title{Unbold if clicked}
\usage{
unbold_if_clicked(title, clicked)
}
\arguments{
\item{title}{A string}

\item{clicked}{A boolean}
}
\description{
Or, really, bold if clicked
}
