% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas_extract_table.R
\name{use_labels_as_headers}
\alias{use_labels_as_headers}
\title{Use labels as headers for a data frame}
\usage{
use_labels_as_headers(t)
}
\arguments{
\item{t}{A data frame}
}
\description{
Uses the labels as the column names for a data frame
}
