% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_err_width.R
\name{v_err_width}
\alias{v_err_width}
\title{Get Error Width (using T-test)}
\usage{
v_err_width(x, ci = 0.95, error_calc = c("Normal", "Percent"))
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Get Error Width (using T-test)
}
