% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_invert.R
\name{v_invert}
\alias{v_invert}
\title{Invert a named vector}
\usage{
v_invert(x)
}
\arguments{
\item{x}{A named vector}
}
\description{
Sets the names of the vector to the value, and the values as the names
}
\examples{
v_invert(c("a" = "A", "b" = "B"))

}
