% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_extract_table.R
\name{v_replace_empty_with_tmps}
\alias{v_replace_empty_with_tmps}
\title{Replace empty with tmp}
\usage{
v_replace_empty_with_tmps(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector
}
\description{
Replace empty with tmp
}
