% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_revalue.R
\name{v_revalue}
\alias{v_revalue}
\title{Revalue a vector}
\usage{
v_revalue(x, r)
}
\arguments{
\item{x}{a Vector}

\item{r}{a named vector where the names are what to replace the values with}
}
\description{
Replace values in a vector
}
\examples{
revalue(x = c("a","b","c"), r = c("x"="a", "y"="b"))

}
