% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_extract_table.R
\name{v_rtf_format}
\alias{v_rtf_format}
\title{Separate and title a data frame with character vector \code{value}}
\usage{
v_rtf_format(s)
}
\arguments{
\item{s}{A vector of strings}
}
\description{
Sets the names of the data frame and separates the \code{value} column into columns. Assumes the titles are
in the first row
}
